move consign legacy analytics to related components
