+++ src/desktop/assets/analytics.ts
@@ -47,7 +47,6 @@ $(() =>
     require("../analytics/articles.js")
     require("../analytics/artists.js")
     require("../analytics/artwork.js")
-    require("../analytics/consignments.js")
     require("../analytics/contact.js")
     require("../analytics/embedded_inquiry.js")
     require("../analytics/galleries.js")
